Move SDK common files to Foundation common directory
