+++ scripts/foundation/templates/entities.group.entity.jinja2
@@ -9,9 +9,9 @@ from __future__ import unicode_literals
 from builtins import str  # noqa
 from builtins import super
 
-from mbed_cloud.sdk.common.entity import Entity
-from mbed_cloud.sdk.common import fields
-from mbed_cloud.sdk import enums
+from mbed_cloud.foundation.common.entity_base import Entity
+from mbed_cloud.foundation.common import fields
+from mbed_cloud.foundation import enums
 
 {% for entity in entities %}
 
@@ -80,7 +80,7 @@ class {{ entity._key|to_pascal }}(Entity):
         :rtype: {{ field.python_type }}{% if field['items'] and field['items'].foreign_key %}[{{ field['items'].foreign_key.entity|to_pascal }}]{% endif %}{% if field.foreign_key %}[{{ field.foreign_key.entity|to_pascal }}]{% endif %}
         """
         {% if field.getter_custom_method %}
-        from mbed_cloud.sdk.common._custom_methods import {{ field.getter_custom_method }}
+        from mbed_cloud.foundation._custom_methods import {{ field.getter_custom_method }}
         return {{ field.getter_custom_method }}(self=self, field=self._{{field._key|to_snake}})
         {% else %}
         return self._{{field._key|to_snake}}.value
@@ -96,7 +96,7 @@ class {{ entity._key|to_pascal }}(Entity):
         :type value: {{ field.python_type }}{% if field['items'] and field['items'].foreign_key %}[{{ field['items'].foreign_key.entity|to_pascal }}]{% endif %}{% if field.foreign_key %}[{{ field.foreign_key.entity|to_pascal }}]{% endif %}
         """
         {% if field.setter_custom_method %}
-        from mbed_cloud.sdk.common._custom_methods import {{ field.setter_custom_method }}
+        from mbed_cloud.foundation._custom_methods import {{ field.setter_custom_method }}
         {{ field.setter_custom_method }}(self=self, field=self._{{field._key|to_snake}}, value=value)
         {% else %}
         self._{{field._key|to_snake}}.set(value)
@@ -153,13 +153,13 @@ class {{ entity._key|to_pascal }}(Entity):
             {%- set field_name = field._key|to_snake %}
         {%- endif %}
         :param {{field_name}}: {{ (field.description or field.foreign_key.entity|to_pascal)|wordwrap(width=70)|indent(width=12) }}
-        :type {{field_name}}: {{ (field.foreign_key and 'mbed_cloud.sdk.' + field.foreign_key.entity|to_pascal) or field.python_type}}
+        :type {{field_name}}: {{ (field.foreign_key and 'mbed_cloud.foundation.entities.' + field.foreign_key.entity|to_pascal) or field.python_type}}
         {% endif %}{% endfor %}
         :return: An iterator object which yields instances of an entity.
         {%- else %}
         {% for field in method.fields %}{% if field.external_param %}
         :param {{field._key|to_snake}}: {{ (field.description or field.foreign_key.entity|to_pascal)|wordwrap(width=70)|indent(width=12) }}
-        :type {{field._key|to_snake}}: {{ (field.foreign_key and 'mbed_cloud.sdk.' + field.foreign_key.entity|to_pascal) or field.python_type}}
+        :type {{field._key|to_snake}}: {{ (field.foreign_key and 'mbed_cloud.foundation.entities.' + field.foreign_key.entity|to_pascal) or field.python_type}}
         {% endif %}{% endfor %}
         {%- endif %}
         :rtype: {{ return_type }}
@@ -174,8 +174,8 @@ class {{ entity._key|to_pascal }}(Entity):
     {% if method.defer_to_foreign_key_field %}
         return {{ method.defer_to_foreign_key_field.field }}.{{ method.defer_to_foreign_key_field.method }}()
     {% elif method.custom_method %}
-        from mbed_cloud.sdk.common._custom_methods import {{ method.custom_method }}
-        {% if method.foreign_key %}from mbed_cloud.sdk.entities import {{ method.foreign_key.entity|to_pascal }}{% endif %}
+        from mbed_cloud.foundation._custom_methods import {{ method.custom_method }}
+        {% if method.foreign_key %}from mbed_cloud.foundation import {{ method.foreign_key.entity|to_pascal }}{% endif %}
 
         return {{ method.custom_method }}(
             self=self,
@@ -195,7 +195,7 @@ class {{ entity._key|to_pascal }}(Entity):
             {% if method.pagination %},wraps=self._{{ method.paginate_target }}{% endif %}
         )
     {% else %}
-        {% if foreign_key_import %}from mbed_cloud.sdk.entities import {{ return_type }}{% endif %}
+        {% if foreign_key_import %}from mbed_cloud.foundation import {{ return_type }}{% endif %}
         return self._client.call_api(
             method='{{method.method}}',
             path='{{ method.path }}',
