+++ src/mbed_cloud/foundation/common/enum_base.py
