+++ src/mbed_cloud/foundation/entities/accounts/account.py
@@ -9,9 +9,9 @@ from __future__ import unicode_literals
 from builtins import str  # noqa
 from builtins import super
 
-from mbed_cloud.sdk.common.entity import Entity
-from mbed_cloud.sdk.common import fields
-from mbed_cloud.sdk import enums
+from mbed_cloud.foundation.common.entity_base import Entity
+from mbed_cloud.foundation.common import fields
+from mbed_cloud.foundation import enums
 
 
 class Account(Entity):
@@ -1263,7 +1263,7 @@ class Account(Entity):
         :rtype: mbed_cloud.pagination.PaginatedResponse(Account)
         """
 
-        from mbed_cloud.sdk.common._custom_methods import paginate
+        from mbed_cloud.foundation._custom_methods import paginate
 
         return paginate(
             self=self,
@@ -1525,8 +1525,8 @@ class Account(Entity):
         :rtype: mbed_cloud.pagination.PaginatedResponse(SubtenantTrustedCertificate)
         """
 
-        from mbed_cloud.sdk.common._custom_methods import paginate
-        from mbed_cloud.sdk.entities import SubtenantTrustedCertificate
+        from mbed_cloud.foundation._custom_methods import paginate
+        from mbed_cloud.foundation import SubtenantTrustedCertificate
 
         return paginate(
             self=self,
@@ -1601,8 +1601,8 @@ class Account(Entity):
         :rtype: mbed_cloud.pagination.PaginatedResponse(SubtenantUserInvitation)
         """
 
-        from mbed_cloud.sdk.common._custom_methods import paginate
-        from mbed_cloud.sdk.entities import SubtenantUserInvitation
+        from mbed_cloud.foundation._custom_methods import paginate
+        from mbed_cloud.foundation import SubtenantUserInvitation
 
         return paginate(
             self=self,
@@ -1638,8 +1638,8 @@ class Account(Entity):
         :rtype: mbed_cloud.pagination.PaginatedResponse(SubtenantUser)
         """
 
-        from mbed_cloud.sdk.common._custom_methods import paginate
-        from mbed_cloud.sdk.entities import SubtenantUser
+        from mbed_cloud.foundation._custom_methods import paginate
+        from mbed_cloud.foundation import SubtenantUser
 
         return paginate(
             self=self,
