+++ src/mbed_cloud/foundation/entities/accounts/user.py
@@ -9,9 +9,9 @@ from __future__ import unicode_literals
 from builtins import str  # noqa
 from builtins import super
 
-from mbed_cloud.sdk.common.entity import Entity
-from mbed_cloud.sdk.common import fields
-from mbed_cloud.sdk import enums
+from mbed_cloud.foundation.common.entity_base import Entity
+from mbed_cloud.foundation.common import fields
+from mbed_cloud.foundation import enums
 
 
 class User(Entity):
@@ -744,8 +744,8 @@ class User(Entity):
         :rtype: mbed_cloud.pagination.PaginatedResponse(User)
         """
 
-        from mbed_cloud.sdk.common._custom_methods import paginate
-        from mbed_cloud.sdk.entities import User
+        from mbed_cloud.foundation._custom_methods import paginate
+        from mbed_cloud.foundation import User
 
         return paginate(
             self=self,
