+++ src/mbed_cloud/foundation/entities/devices/device_enrollment.py
@@ -9,9 +9,9 @@ from __future__ import unicode_literals
 from builtins import str  # noqa
 from builtins import super
 
-from mbed_cloud.sdk.common.entity import Entity
-from mbed_cloud.sdk.common import fields
-from mbed_cloud.sdk import enums
+from mbed_cloud.foundation.common.entity_base import Entity
+from mbed_cloud.foundation.common import fields
+from mbed_cloud.foundation import enums
 
 
 class DeviceEnrollment(Entity):
@@ -286,8 +286,8 @@ class DeviceEnrollment(Entity):
         :rtype: mbed_cloud.pagination.PaginatedResponse(DeviceEnrollment)
         """
 
-        from mbed_cloud.sdk.common._custom_methods import paginate
-        from mbed_cloud.sdk.entities import DeviceEnrollment
+        from mbed_cloud.foundation._custom_methods import paginate
+        from mbed_cloud.foundation import DeviceEnrollment
 
         return paginate(
             self=self,
