+++ src/mbed_cloud/foundation/entities/devices/device_events.py
@@ -9,9 +9,9 @@ from __future__ import unicode_literals
 from builtins import str  # noqa
 from builtins import super
 
-from mbed_cloud.sdk.common.entity import Entity
-from mbed_cloud.sdk.common import fields
-from mbed_cloud.sdk import enums
+from mbed_cloud.foundation.common.entity_base import Entity
+from mbed_cloud.foundation.common import fields
+from mbed_cloud.foundation import enums
 
 
 class DeviceEvents(Entity):
@@ -355,8 +355,8 @@ class DeviceEvents(Entity):
         :rtype: mbed_cloud.pagination.PaginatedResponse(DeviceEvents)
         """
 
-        from mbed_cloud.sdk.common._custom_methods import paginate
-        from mbed_cloud.sdk.entities import DeviceEvents
+        from mbed_cloud.foundation._custom_methods import paginate
+        from mbed_cloud.foundation import DeviceEvents
 
         return paginate(
             self=self,
