+++ src/mbed_cloud/foundation/entities/security/developer_certificate.py
@@ -9,9 +9,9 @@ from __future__ import unicode_literals
 from builtins import str  # noqa
 from builtins import super
 
-from mbed_cloud.sdk.common.entity import Entity
-from mbed_cloud.sdk.common import fields
-from mbed_cloud.sdk import enums
+from mbed_cloud.foundation.common.entity_base import Entity
+from mbed_cloud.foundation.common import fields
+from mbed_cloud.foundation import enums
 
 
 class DeveloperCertificate(Entity):
@@ -266,7 +266,7 @@ class DeveloperCertificate(Entity):
         :rtype: TrustedCertificate
         """
 
-        from mbed_cloud.sdk.entities import TrustedCertificate
+        from mbed_cloud.foundation import TrustedCertificate
 
         return self._client.call_api(
             method="get",
