+++ src/mbed_cloud/foundation/entities/security/enums.py
@@ -9,7 +9,7 @@ from __future__ import unicode_literals
 from builtins import str  # noqa
 from builtins import object
 
-from mbed_cloud.sdk.common.enum import BaseEnum
+from mbed_cloud.foundation.common.enum_base import BaseEnum
 
 
 class CertificateEnrollmentEnrollResultEnum(BaseEnum):
