+++ src/mbed_cloud/foundation/entities/security/trusted_certificate.py
@@ -9,9 +9,9 @@ from __future__ import unicode_literals
 from builtins import str  # noqa
 from builtins import super
 
-from mbed_cloud.sdk.common.entity import Entity
-from mbed_cloud.sdk.common import fields
-from mbed_cloud.sdk import enums
+from mbed_cloud.foundation.common.entity_base import Entity
+from mbed_cloud.foundation.common import fields
+from mbed_cloud.foundation import enums
 
 
 class TrustedCertificate(Entity):
@@ -299,7 +299,7 @@ class TrustedCertificate(Entity):
         :rtype: bool
         """
 
-        from mbed_cloud.sdk.common._custom_methods import (
+        from mbed_cloud.foundation._custom_methods import (
             is_developer_certificate_getter,
         )
 
@@ -315,7 +315,7 @@ class TrustedCertificate(Entity):
         :type value: bool
         """
 
-        from mbed_cloud.sdk.common._custom_methods import (
+        from mbed_cloud.foundation._custom_methods import (
             is_developer_certificate_setter,
         )
 
@@ -562,7 +562,7 @@ class TrustedCertificate(Entity):
         :rtype: DeveloperCertificate
         """
 
-        from mbed_cloud.sdk.entities import DeveloperCertificate
+        from mbed_cloud.foundation import DeveloperCertificate
 
         return self._client.call_api(
             method="get",
@@ -595,8 +595,8 @@ class TrustedCertificate(Entity):
         :rtype: mbed_cloud.pagination.PaginatedResponse(TrustedCertificate)
         """
 
-        from mbed_cloud.sdk.common._custom_methods import paginate
-        from mbed_cloud.sdk.entities import TrustedCertificate
+        from mbed_cloud.foundation._custom_methods import paginate
+        from mbed_cloud.foundation import TrustedCertificate
 
         return paginate(
             self=self,
