--- src/mbed_cloud/sdk/common/_custom_methods/__init__.py
@@ -1,9 +1,0 @@
-from mbed_cloud.sdk.common._custom_methods.device import (
-    download_full_report_file,
-    download_errors_report_file,
-)
-from mbed_cloud.sdk.common._custom_methods.security import (
-    is_developer_certificate_getter,
-    is_developer_certificate_setter,
-)
-from mbed_cloud.sdk.common._custom_methods.common import paginate
