+++ tests/unit/foundation/test_entity.py
@@ -6,8 +6,8 @@ import datetime
 
 from tests.common import BaseCase
 
-from mbed_cloud.sdk.common.entity import Entity
-from mbed_cloud.sdk.common import fields
+from mbed_cloud.foundation.common.entity_base import Entity
+from mbed_cloud.foundation.common import fields
 
 
 class SimpleEntity(Entity):
